native: calcola il modulo del peso a 64 bit, std::abs(INT_MIN) e' UB
effect.bias e' un int che arriva da evaluate_arithmetic_expression, dove
niente limita l'overflow: se un __weight(...) vale esattamente INT_MIN,
std::abs(INT_MIN) e' comportamento indefinito perche' -INT_MIN non e'
rappresentabile come int. Il ramo Sign lo chiamava dal commit precedente.

Il modulo si calcola ora in long long, dove -INT_MIN c'e', tramite un
helper totale accanto a normalize_sign; rank_weight diventa long long in
ResolvedModifierValue e nella firma di update_best_by_weight. Restano
invariati il valore memorizzato (la direzione +1/-1) e i rami Level, True
e False, che continuano a ordinare per bias.

Il test nuovo fa gareggiare una direttiva sign a INT_MIN contro una a
INT_MAX: vincere richiede 2147483648 > 2147483647, cioe' un modulo davvero
calcolato a 64 bit.

+++ clingo-native/libclingo/clingo/heuristic_propagator.hh
@@ -41,10 +41,11 @@ private:
     // rank_weight e' il criterio di ordinamento, cioe' la forza del candidato;
     // value e' cio' che lo slot memorizza. Per `level` i due coincidono col
     // peso, per `sign` no: li' value e' la direzione (+1/-1) e rank_weight e'
-    // il modulo del peso.
+    // il modulo del peso. E' un long long perche' il modulo di INT_MIN non e'
+    // rappresentabile come int.
     struct ResolvedModifierValue {
         bool active = false;
-        int rank_weight = 0;
+        long long rank_weight = 0;
         int value = 0;
         size_t source_candidate_id = 0;
     };
@@ -186,7 +187,7 @@ private:
                                    Clingo::Assignment const &assignment,
                                    CandidateHeuristicEffect &effect) const;
     void update_best_by_weight(ResolvedModifierValue &current,
-                               int rank_weight,
+                               long long rank_weight,
                                int value,
                                size_t candidate_id) const;
     void apply_effect_to_target_state(CandidateHeuristicEffect const &effect,
